requirements.txt
fastapi
fastmcp
uvicorn
PyJWT>=2.8.0
cryptography>=41.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0